+++ new_software/microcontroller/TipPositionLog.hpp
@@ -5,7 +5,7 @@ class TipPositionLog {
     uint8_t x;
     uint8_t y;
     uint16_t z; // piezo position
-    float currentSignal; // V
+    uint16_t currentSignal; // 0xffff/3.3 V
   };
 
   static const int size_ = 100;
@@ -15,5 +15,5 @@ class TipPositionLog {
 
 public:
   void flush();
-  void add(uint8_t, uint8_t, uint16_t, float);
+  void add(uint8_t, uint8_t, uint16_t, uint16_t);
 };
